UnlockApp is now a functional component
